refactor: streamline uploads-page.vue by modularizing components, enhancing event handling, and improving layout for better user experience
