+++ src/renderer/src/utils/image.ts
@@ -1,0 +1,77 @@
+/**
+ * Image processing utilities extracted from uploads-page.vue
+ */
+
+export async function resizePreview(file: File, maxWidth: number): Promise<string> {
+  // Use data URL to satisfy CSP (img-src 'self' data:)
+  const dataUrlSrc = await new Promise<string>((resolve, reject) => {
+    const reader = new FileReader()
+    reader.onload = () => resolve(reader.result as string)
+    reader.onerror = (e) => reject(e)
+    reader.readAsDataURL(file)
+  })
+  try {
+    const img = await loadImage(dataUrlSrc)
+    const scale = Math.min(1, maxWidth / (img.width || 1))
+    const w = Math.max(1, Math.round((img.width || maxWidth) * scale))
+    const h = Math.max(1, Math.round((img.height || maxWidth) * scale))
+    const canvas = document.createElement('canvas')
+    canvas.width = w
+    canvas.height = h
+    const ctx = canvas.getContext('2d')!
+    ctx.drawImage(img, 0, 0, w, h)
+    const dataUrl = canvas.toDataURL(file.type || 'image/png', 0.8)
+    return dataUrl.split(',')[1] || ''
+  } finally {
+    // nothing to revoke when using FileReader data URL
+  }
+}
+
+export function loadImage(url: string): Promise<HTMLImageElement> {
+  return new Promise((resolve, reject) => {
+    const img = new Image()
+    img.onload = () => resolve(img)
+    img.onerror = (e) => reject(e)
+    img.src = url
+  })
+}
+
+export function fileToBase64(file: File): Promise<string> {
+  return new Promise((resolve, reject) => {
+    const reader = new FileReader()
+    reader.onload = () => {
+      const result = reader.result as string
+      const comma = result.indexOf(',')
+      resolve(comma >= 0 ? result.slice(comma + 1) : result)
+    }
+    reader.onerror = (e) => reject(e)
+    reader.readAsDataURL(file)
+  })
+}
+
+export function dataUrl(mime: string, base64?: string | null) {
+  if (!base64) return ''
+  return `data:${mime};base64,${base64}`
+}
+
+export function formatFileSize(bytes: number): string {
+  const units = ['B', 'KB', 'MB', 'GB']
+  let size = bytes
+  let unitIndex = 0
+  
+  while (size >= 1024 && unitIndex < units.length - 1) {
+    size /= 1024
+    unitIndex++
+  }
+  
+  return `${size.toFixed(unitIndex > 0 ? 1 : 0)} ${units[unitIndex]}`
+}
+
+export function formatDate(timestamp: number): string {
+  return new Date(timestamp * 1000).toLocaleDateString('zh-CN', {
+    month: 'short',
+    day: 'numeric',
+    hour: '2-digit',
+    minute: '2-digit'
+  })
+}
